+++ src/components/Currency/Currency.styled.js
